Cleaup structure
